Refactored library and tests
So I'm a pedant, and it was really bothering me that the test was testing
KingDiamond and not the track Melissa.  Considering that the track Melissa could
be a model and burning Melissa was business logic, I refactored the test along
those lines.  This is actually an integration or functional test, but for the
purposes of lulz it's close enough.